requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

torch>=2.1.0
transformers>=4.36.0

google-api-python-client==2.118.0
python-dotenv==1.0.1

regex
requests
